requirements.txt
jax
matplotlib
numpy
scipy
pyyaml
dotmap
setuptools
pillow
